refactor: split attribution exposure points
