+++ quality/baseline_report.md
@@ -13,11 +13,11 @@ completion claim.
 ## Generation Identity
 
 - Git branch: `feat/enterprise-risk-refactor-continuation`
-- Git commit: `ed8b3b97442c8ce26366f31f3565eea7440a348c`
+- Git commit: `5141005749ae60b155ada8e1d285bf2a966d10c9`
 
 ## Current Code Size
 
-- Python source files under `src/`: 219
+- Python source files under `src/`: 220
 - Python test files under `tests/`: 103
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -26,7 +26,6 @@ completion claim.
 
 | Path | Lines | Bytes |
 | --- | --- | --- |
-| src/app/services/attribution_exposure_history.py | 280 | 9754 |
 | src/app/services/attribution_decomposition.py | 279 | 9057 |
 | src/app/services/attribution_stateful_inputs.py | 276 | 9699 |
 | src/app/openapi_examples.py | 273 | 8965 |
@@ -46,6 +45,7 @@ completion claim.
 | src/app/services/concentration/parsing.py | 231 | 7793 |
 | src/app/services/concentration/resolvers.py | 231 | 7566 |
 | src/app/services/rolling_metric_series.py | 221 | 6803 |
+| src/app/services/benchmark_exposure_history.py | 220 | 7264 |
 
 ### Largest Functions And Classes
 
@@ -96,12 +96,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-337 files already formatted
+338 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 322 source files
+Success: no issues found in 323 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -115,10 +115,10 @@ Success: no issues found in 322 source files
 ...
 src\app\services\rolling_stateful_source_responses.py         45      1     16      1    97%   96
 ------------------------------------------------------------------------------------------------------
-TOTAL                                                       6095     91   1028     73    98%
+TOTAL                                                       6103     91   1028     73    98%
 
-176 files skipped due to complete coverage.
-550 passed in 8.71s
+177 files skipped due to complete coverage.
+551 passed in 9.22s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -148,7 +148,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 219 files...
+Scanning 220 files...
 
 Success! No dependency issues found.
 ```
@@ -239,15 +239,15 @@ tests/unit/test_agent_effectiveness_review.py::test_agent_effectiveness_review_r
 tests/unit/test_agent_effectiveness_review.py::test_codebase_review_playbook_requires_recurring_effectiveness_review
 tests/unit/test_agent_effectiveness_review.py::test_second_effectiveness_review_records_configuration_and_context_improvements
 tests/unit/test_agent_effectiveness_review.py::test_third_effectiveness_review_records_problem_details_and_modularity_evidence
+tests/unit/test_agent_effectiveness_review.py::test_fourth_effectiveness_review_records_recent_modularity_evidence
 tests/unit/test_app_factory.py::test_create_app_builds_independent_service_instance
-tests/unit/test_app_factory.py::test_create_app_registers_risk_analytics_routes
 ...
 tests/unit/test_upstream_errors.py::test_classify_upstream_http_error_matrix[504-502-UPSTREAM_FAILURE-upstream_failure-True]
 tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matrix[exc0-504-UPSTREAM_TIMEOUT-timeout]
 tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matrix[exc1-503-UPSTREAM_UNAVAILABLE-transport]
 tests/unit/test_upstream_errors.py::test_invalid_payload_and_missing_data_carry_structured_categories
 
-550 tests collected in 1.58s
+551 tests collected in 1.72s
 ```
 - Import-linter report-only: passed
